requirements.txt
anyio==3.7.0
axios==0.4.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
numpy==1.24.3
pandas==2.0.2
psycopg2-binary==2.9.6
pydantic==1.10.9
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
requests==2.31.0
rich==13.4.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
zope.interface==6.0